Change delta_time precision to double

+++ src/tileset.cpp
@@ -133,7 +133,7 @@ Tile Tileset::GetTileFromId(unsigned int tile_id) {
 	return tiles_[tile_id];
 }
 
-void Tileset::Update(int delta_time) {
+void Tileset::Update(double delta_time) {
 	for (std::pair<unsigned int, Tile> tile_pair : tiles_) {
 		tile_pair.second.Update(delta_time);
 	}
